Add comments and rename some things
