+++ frontend/src/App.js
@@ -2,7 +2,7 @@ import { Component } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./pages/Home.js";
 import JoinGame from "./pages/JoinGame.js";
-import Game from "./pages/Game.js";
+import Game from "./pages/GamePage.js";
 import RegisterPage from "./pages/RegisterPage.js";
 import SignInPage from "./pages/SignInPage.js";
 import SignOutPage from "./pages/SignOutPage.js";
